+++ app/collections/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Metadata } from "next";
 import { CollectionsPage } from "../src/components/pages/collectionsPage/collectionsPage";
-import { getCollectionsData } from "@/src/utils/apiHelpers";
+import { CollectionsApi } from "@/src/utils/apiClients";
 import { redirect } from "next/navigation";
 import CollectionSearchParams from "@/src/types/CollectionSearchParams";
 import PageLayout from "@/src/components/pageLayout/pageLayout";
@@ -19,7 +19,7 @@ export const metadata: Metadata = {
 };
 
 export default async function Collections({ searchParams }: CollectionsProps) {
-  const data = await getCollectionsData({
+  const data = await CollectionsApi.getCollectionsData({
     keyword: searchParams.q,
     sort: searchParams.sort,
     page: searchParams.page,
@@ -42,7 +42,7 @@ export default async function Collections({ searchParams }: CollectionsProps) {
       ]}
       adobeAnalyticsPageName={createAdobeAnalyticsPageName("all-collections")}
     >
-      <CollectionsPage collectionSearchParams={searchParams} data={data} />
+      <CollectionsPage collectionsSearchParams={searchParams} data={data} />
     </PageLayout>
   );
 }