+++ CHANGELOG.md
@@ -10,12 +10,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added custom select component, `selectFilter`, and corresponding modal for facet filters (DR-3394, DR-3396)
+- Added active filters panel, `activeFilters` (DR-3367)
+- Added sort component, `sortMenu` (DR-3366)
 - Added search result model, mock data, card component, and card grid component (DR-3363)
+<<<<<<< HEAD
 - Added custom toggle tip component for rights filter (DR-3460)
+=======
+- Added collection structure component (DR-3415)
+>>>>>>> main
 
 ### Updated
 
 - Updated collection search to use `q` param and added redirect for `collection_keywords` (DR-3337)
+- Updated homepage collections item counts and `/collections` page to use Collections API (DR-3453)
 
 ## [0.3.2] 2025-02-13